Revamp grammar, parser, bump version to 0.6.0-pre.
This uses an updated version of the parser generated
from the latest version of glop.

Stylistically, the major difference is that the
generated code doesn't rely on nested functions;
it also uses lambdas and more helper methods to
reduce the number of lines of code, making the file
2.5x smaller.

The grammar change is a no-op to the language it
recognizes, but by switching from recursion to
iteration on the array and member lists, this
prevents a large class of potential stack overflows
in the python interpreter, and also speeds things
up.

The latest version of glop can optionally produce
a memoizing parser, but that appears to be unneeded
here and actually slows things down.

The current version is 100x - 150x slower than the
pure python version of the equivalent json parser;
this appears to be 1.3 to 2.5x faster than v0.5.1.

(Memoizing made it ~20% slower).