Simplify json5.dump() and json5.dumps().
I don't remember at this  point why I added the `compact`
and `as_json` kwargs to .dump() and .dumps(). I'm guessing it was
some combination of being lazy and/or being conservative in
implementation. The net result of it, though, was that unless
you specified compact=True, you got json-formatted output, not
json5-formatted output. And that now seems wrong.

[ I suppose it's possible that I wanted compact=False by default
in order to get the speedup from json's native implementation, though. ].

This CL removes the compact keyword, and changes the implementation
to always print json5-formatted output (as if you always specified
compact=True).

It also removes the 'as_json' keyword; if you want json-formatted
output, just use json.dump() instead of json5.dump().

json5.tool keeps support for the --as-json keyword, though, because
that allows you to use json5.tool to convert json5 to json, and
that still seems useful.

Also, this fixes a bug I discovered when I turned compact=True by
default -- :) -- which is that in Python, apparently bools are subclasses
of ints, and the constants `False` and `True` have integer values 0 and 1
respectively. This means that '0 == False' is true. In order to
test if something is a bool, you have to use `is` or check the type
explicitly.