Add quote_keys, trailing_commas to dump()/dumps().
This adds support for two new keyword arguments to dump() and dumps().

By default, if an object key is a valid identifier, it will not be
quoted in the output. If you pass quote_keys=True, it will be quoted.

Also, by default, if an object or an array is printed across multiple
lines (i.e., if indent >= 0), then a trailing comma will be added
after the last item. If you pass trailing_commas=False, it will not
be added.

Calling json5.dumps(obj, quote_keys=True, trailing_commas=False) should
produce the same output as calling json.dumps(obj).

This CL also adds the --quote-keys and --no-trailing-commas arguments
to json5.tool, and passing --as-json is now implemented by doing those
two things.