Make checking for duplicate keys possible.
This adds an optional `allow_duplicate_keys` parameter to load()
and loads(). The parameter is True by default, but, if set to False
and an object contains duplicate keys, a ValueError will be raised.

Duplicate key checking is off by default for compatibility with
json.load(), earlier versions of json5.load(), and because it's not
clear if people actually want duplicate key checking enabled by default.

This addresses https://github.com/dpranke/pyjson5/issues/24.