Fix bad error reporting on partial matches.
This patch addresses an error reported in
https://github.com/dpranke/pyjson5/issues/26, where calling
`json.loads('[ ,]') would report an error on column 4, not column 3.

The reason for this had to do with the way the optional whitespace (sp*)
rule was coded in the grammar, and a bug in the underlying parser generator
(glop), where if you had a rule like (a ?())* and you matched a and not b,
the errposition wouldn't be properly rolled back before the conditional
expression. That bug was fixed in glop v0.6.2; this CL regenerates the
JSON5 parser with that version of the parser generator, and then adds
json5-specific tests to check for the correct position reporting.