Add allow_duplicate_keys=False to dump()/dumps().
Commit 46ca46 fixed an issue with dump()/dumps() but introduced
the possibility for the routine to generate ill-formed JSON.

This change adds an optional `allow_duplicate_keys` keyword parameter
to the routines. By default the value is True, and duplicates will
be allowed, to match the behavior of json.dump() and json.dumps().
However, if you pass `allow_duplicate_keys=False`, any duplicates that
result from the string conversions will produce a ValueError exception
instead.