Fix stray trailing comma when dumping an object with an invalid key.
json5.dumps({"foo": 1, (1, 2): 2}, skip_keys=true) would incorrectly print
a trailing comma if the `(1, 2)` item was found as the second item in
the list.

This was reported in https://github.com/dpranke/pyjson5/issues/33. This
only showed up in Python3 and not Python2, because of differences in the
ordering of the dictionary keys. This change also rewrites the test for
this to use OrderedDicts() to address that, and adds a test for a dict
with only invalid keys as well, for comprehensiveness.