Fix ZeroDivisionError in benchmarking code.
Due to some sloppy coding on my part, if a particular benchmark
ran too fast in JSON (most likely on Windows, which has a lower
time.time() resolution), we would divide by zero when computing
the speedup.

This patch rewrites things to avoid that, and makes the output
a bit more detailed.

This fixes https://github.com/dpranke/pyjson5/issues/42.