Bump version to v0.9.6, update package classifiers.
At this point the library can probably be safely considered stable
and production-suitable, assuming you are aware of and okay with
the performance, so this change bumps the setuptools classifiers
accordingly.

It also marks the package as tested against the latest Python3 versions
(3.8 and 3.9) in addition to 2.7 still being supported.

If this release does not surface any significant regressions, I'll probably
promote it to v1.0.0 as well.

If I do ever get around to either significantly speeding it up or
adding round-trip editing, that'll probably be a 2.0 version.