Remove //.vscode/settings.json .
I think this was accidentally included in 50898a4; at any rate I
didn't notice that this was included, and I don't want to add
VSCode settings (or any other file) without being sure that I
want it.

So, I'm happy to re-add this file in a separately reviewed change
if anyone wants it.