Fix issue #47: syntax error points to wrong position ...
Due to the way the parser was handing nested parts of the JSON5
grammar, we would incorrectly report errors as having happened
at the start of a block of content, rather than where the error
actually happened.

A simple case of this would be reporting `"foo"bar` as happening
at column 1 rather than column 5. The bug report (#47) has a more
complicated and realistic example.

To fix this we needed to rev the underlying parser generator (glop)
to v0.6.4 and also rewrite the JSON5 grammar to use a positive
lookahead.