Remove reference to additional_target_cpus
This parameter will be removed from Chromium build system. While
this is just a meaningless key in a sample json file used for
performance testing, the project is cloned into Chromium and the
string "additional_target_cpus" shows up when doing search in
the codebase of Chromium.

Removing the line from the sample should change too much the
validity of the performance testing, but it would simplify the
effort of removing the additional_target_cpus parameter from
Chromium build system as this would remove false positive when
searching the codebase.

Reference: https://crbug.com/1337780