Add typing information.
This change takes a stab at adding typing information to the
module, in order to address #issue65.

Ideally we'd use the same typing as the builtin json module. But,
the json module has no typing information, so I've rolled my
own recursive type in `json5.ObjectType`, and the `dump` routines
take that type as input for the `obj` parameter.

Note that due to the multiple hooks the library provides, I can't
actually say much about what the `load` routines return and have
to use `Any`.