Add `strict` option to load()/loads()
This adds support for the `strict` flag to load() and loads().
By default, strict is True and the behavior is unchanged from
earlier json5 versions. If you pass `strict=False` then it
will allow control characters (\x00 - \x1f) in strings, to
mirror the python `json` package's `strict` flag.

This also adds the `--strict`/`--no-strict` flags to the tool
as well.

This is implemented by modifying the underlying parser generator
to accept "global" variables that can be passed to the parser
to control how the parser operates. (You can do other things
w/globals as well, but this grammar only does the one thing in
this case). As part of this I shuffled around the code in the
glop repo so that the code on `main` is what is used now, instead
of a strange old dev branch version.