Add custom encoding to dump/dumps. (#88)
Add support for custom encoders (and bump version to 0.10.0).

This change reimplements `dump()` and `dumps()` on top of a  `JSON5Encoder` class that can be subclassed to customize the encoding of a data structure. The new class is somewhat similar to the `JSONEncoder` class in the standard library, except that in addition to a `default()` method that you can define to say how to encode an object that the normal encoder can't handle, there's also an `encode()` method that you can override to customize how existing data types are handled.

The change also adds a `quote_style` argument that can be used to customize how strings are encoded (whether to use double or single quotes by default, etc.).