Add a `parse` method.
This adds a public `parse` method that exposes the position
information from the internal `parse` method.

It also adds a `start` parameter than can specify an offset into
a string to start at, and a `consume_trailing` parameter that, if
False, tells the parse to stop reading when a valid object has
been parsed.