Introduce a new `parse` method.
This introduces a new `parse` method that can be used to iterate through
a string, extracting multiple values.

As part of that, it adds a new `consume_trailing` parameter to
`load`/`loads`/`parse` that specifies whether to keep parsing after a
valid object is reached. By default, this is True and the string must
only contain trailing whitespace. If set to False, parsing will stop
when a valid object is reached.

It also adds a new `start` parameter to `load`/`loads`/`parse` to
specify the zero-based offset to start parsing the string or file from.